Switch role-hint for jenkins-module to jm.
This way, we can use <packaging>jm</packaging>, so that the packaging
in the POM matches the actual artifact file extension. This will get
us around a problem where Gradle looks at the POM's "packaging" field
to determine the file extension for an artifact. This means that if
you end up with a dependency on something with packaging
"jenkins-module", it expects to find a file ending with
".jenkins-module" and fails otherwise. So we really want the role-hint
(which is what you specify in <packaging>) to match the file
extension, I think.

That said, I'll also ask Hans Dockter if he has any thoughts. =)