Fixed the discrepancy between the hpl file and plugins staged.
Talking to Max wrt Jenkow plugin revealed this problem.

As the code was written, the "compile" scope was used to place plugins
to work/plugins, while the "test" scope was used to generate a list of
dependencies in the generated .hpl file for the current plugin.

With this, we end up producing more in the "Dependencies" field of the
hpl file than what's available in work/plugins, so at the start-up the
plugin currently developed will fail to load, citing missing
dependencies.

The hpl generation code needs to use the same dependency scope as
the RunMojo to avoid this.