[JENKINS-41196] don't put core classes into WebAppClassLoader
With JENKINS-41196 change, core classes need to be loaded from a
classloader created during bootstrap, not from webappclassloader.

This means if we want to keep the developer experience optimization like
this, the fix needs to be made in the bootstrap.

For the time being, I think people who benefit from this is limited, so
I think it's OK to just remove it.