handle SNAPSHOT builds correctly
if the war scanning encounters a SNAPSHOT build of a plugin then then
manifest file will contain a `Plugin-Version` entry like
`1.2.3-4-SNAPSHOT (private abcd1234-username)`.

This causes a failure when updating the dependencies as the plugin will
try and use that directly without stripping off the extra parts.

THis strips off any extra parts should they be present so that the
version number matches a valid maven version number and also emits a
warning to the log about non repeatable builds if any snapshots are
encountered.