remove Plugin-Module manifest entry
THe manifest entry was not correctly caclulated, being different
depending on where you ran maven from and if you also had any local
aggrgator project or not.

The original intention and need of this was to use the path from the git
repostory for the PCT, but the code changed from using a git path along
the way.

As a result the code produced non deterministic results and the PCT
which needed this information has been updated to not need this in https://github.com/jenkinsci/plugin-compat-tester/pull/557

because the information is non deterministic and there are no consumers
of this, it is simpler to remove the code.