Revert #21 to allow configuration discovery
I added this line 10 years ago to address a start-up delay. However I
need now to remove it in order to allow configuration discovery.

I couldn't notice any significant delay by removing this, so I assume that
the previous delay was caused by a Jetty bug (back in 8.x) that is now
fixed in the Jetty version the Jenkins project uses.