Add property for detection of inline javascript
Add property `jelly.requireNoInlineJS`. This is passed to the injected
tests as parameters.
When https://github.com/jenkinsci/jenkins-test-harness/pull/868 gets
merged this will b enable the checks.
By default the checks are enabled.
At a later point in time we might consider removing this property so
that it is no longer possible to disable the check. Or we might ignore
it when the Jenkins core version is higher than a certain version.